Time: 61 ms (72.58%), Space: 54 MB (50.53%) - LeetHub

+++ 0493-reverse-pairs/0493-reverse-pairs.java
@@ -81,7 +81,7 @@ class Solution {
         int cnt = 0;
         
         for (int i = low; i <= mid; i++) {
-            while (right <= high && (long)arr[i] > (long)(2 * (long)arr[right])){
+            while (right <= high && (long)arr[i] > (2 * (long)arr[right])){
                 right++;
             } 
             cnt += (right - (mid + 1));
